No Logic Changes. Refactored syntax-highlight feature into code shared between new and old uis
