+++ src/syntax-highlight/common/get-extension.spec.js
