+++ src/syntax-highlight/common/language-ext.js
