+++ src/syntax-highlight/common/source-handler.js
@@ -6,15 +6,14 @@ import languagesExtensions from './language-ext'
 declare var Prism: { languages: { [language: string]: Object } }
 
 /**
- * Retrieves a class according to the element data-filename or data-path attribute.
- * For example, calling it passing an element with a data-filename (or data-path)
- * like "/path/to/file/filename.java" would return "language-java".
+ * Retrieves a class according to a filepath string argument.
+ * For example, calling it passing a string such as:
+ * "/path/to/file/filename.java" would return "language-java".
  *
- * @param  {HTMLElement} element An HTML element. Pass anything different and bear the consequences :)
+ * @param  {string} filePath A file name and (optionally) path.  For example C:/HellowWorld.Java or just HelloWorld.Java
  * @return {string} The class extracted from the element's file path.
  */
-export function getLanguageClass(element: HTMLElement): string {
-    const filePath = getFilepathFromElement(element)
+export function getLanguageClass(filePath: string): string {
     const fileExtension = getExtension(filePath).toLowerCase()
 
     if (fileExtension in languagesExtensions) {
@@ -32,21 +31,6 @@ export function getLanguageClass(element: HTMLElement): string {
     }
 
     return ''
-}
-
-/**
- * Retrieves the filename of an element according to its `data-identifier`, * `data-filename` or `data-path` attributes.
- *
- * @param  {HTMLElement} element An HTML element. Pass anything different and bear the consequences :)
- * @return {string} The filename
- */
-export function getFilepathFromElement(element: HTMLElement): string {
-    const filepath =
-        element.getAttribute('data-identifier') ||
-        element.getAttribute('data-filename') ||
-        element.getAttribute('data-path') ||
-        ''
-    return filepath.trim()
 }
 
 /**
